pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "excel-api"
version = "0.1.0"
description = "API for generating Excel files from JSON data"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "pydantic==2.5.0",
    "openpyxl>=3.1.0",
    "python-multipart==0.0.6",
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "pytest==7.4.4",
    "pytest-cov==4.1.0",
    "pytest-asyncio==0.23.2",
    "httpx==0.25.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]